Reorganize GoldenLayout types and add clarifying documentation
- Moved all type definitions from goldenlayout-types.ts to components.interfaces.ts
  to eliminate circular dependencies and reduce import complexity
- Added comprehensive comments explaining:
  - Why fromGoldenLayoutConfig validation is mandatory (not optional)
  - What 'legacy' means in conversion functions and migration timeline
  - Purpose of SerializedLayoutState for Phase 2/3 serialization
  - Security implications of unvalidated configurations
- Fixed inline import syntax that violated coding standards
- Removed now-redundant goldenlayout-types.ts file

This consolidation keeps all component-related types together and makes
the codebase cleaner while maintaining the same functionality.
