+++ static/panes/conformance-view.ts
@@ -32,9 +32,9 @@ import {unwrapString} from '../assert.js';
 import * as BootstrapUtils from '../bootstrap-utils.js';
 import {CompilationStatus} from '../compiler-service.interfaces.js';
 import {CompilerService} from '../compiler-service.js';
+import {createTypedDragSource, legacyComponentConfigToTyped} from '../components.interfaces.js';
 import * as Components from '../components.js';
 import {SourceAndFiles} from '../download-service.js';
-import {createTypedDragSource, legacyComponentConfigToTyped} from '../goldenlayout-types.js';
 import {Hub} from '../hub.js';
 import * as LibUtils from '../lib-utils.js';
 import {SelectedLibraryVersion} from '../libraries/libraries.interfaces.js';
