+++ static/panes/tree.ts
@@ -30,9 +30,9 @@ import _ from 'underscore';
 import {escapeHTML} from '../../shared/common-utils.js';
 import {assert, unwrap, unwrapString} from '../assert.js';
 import * as BootstrapUtils from '../bootstrap-utils.js';
+import {createTypedDragSource, legacyComponentConfigToTyped} from '../components.interfaces.js';
 import * as Components from '../components.js';
 import {EventHub} from '../event-hub.js';
-import {createTypedDragSource, legacyComponentConfigToTyped} from '../goldenlayout-types.js';
 import {Hub} from '../hub.js';
 import {LanguageKey} from '../languages.interfaces.js';
 import {LineColouring} from '../line-colouring.js';
